requirements/base.txt
# Bleeding edge Django
django==1.7.5

# Configuration
django-configurations==0.8
django-secure==1.0.1
django-cache-url==0.8.0
dj-database-url==0.3.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.3.0

# Models
django-model-utils==2.2

# Images
Pillow==2.7.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.19.1

# For the persistence stores
psycopg2==2.6
MySQL-python==1.2.5

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.7.2

# Useful things
django-avatar==2.0

# Time zones support
pytz==2014.10

# custom requirements
rapid-sms==0.19.0
git+git://github.com/nyaruka/smartmin.git@master
celery==3.1.17
django-bootstrap3==5.1.1
django_select2==4.2.2
django-bootstrap3-datetimepicker==2.2.3
django-scheduler==0.7.4

